+++ example/field_manipulation/field_manipulation_main.g.dart
@@ -37,10 +37,10 @@ abstract class _$PlayerJsonSerializer implements Serializer<Player> {
     if (model is! Player) {
       model = createModel();
     }
-    model.name = map["N"];
-    model.email = map["E"];
-    model.age = map["A"];
-    model.score = map["S"];
+    model.name = map["N"] ?? model.name;
+    model.email = map["E"] ?? model.email;
+    model.age = map["A"] ?? model.age;
+    model.score = map["S"] ?? model.score;
     return model;
   }