+++ example/user/book/book_mongo.dart
@@ -9,9 +9,9 @@ export 'book.dart' show Book;
 part 'book_mongo.g.dart';
 
 @GenSerializer(fields: const {
-  'id': const EnDecode('_id'),
-  'name': const EnDecode('N'),
-  'viewSerializer': const Ignore(),
+  'id': const [const EnDecode('_id')],
+  'name': const [const EnDecode('N')],
+  'viewSerializer': const [ignore],
 }, processors: const {
   'id': const MongoId(),
 })