+++ test/serializer/common/models/bibliotek/book.g.dart
@@ -49,13 +49,16 @@ abstract class _$BookSerializer implements Serializer<Book> {
     if (model is! Book) {
       model = createModel();
     }
-    model.name = map["name"];
-    model.tags = map["tags"]?.map((String val) => val)?.toList();
+    model.name = map["name"] ?? model.name;
+    model.tags = map["tags"]?.map((String val) => val)?.toList() ?? model.tags;
     model.publishedDates = publishedDatesMapKeyNumToStringProcessor
-        .deserialize(map["publishedDates"]);
+            .deserialize(map["publishedDates"]) ??
+        model.publishedDates;
     model.authors = map["authors"]
-        ?.map((Map val) => fromAuthorSerializer.fromMap(val, typeKey: typeKey))
-        ?.toList();
+            ?.map((Map val) =>
+                fromAuthorSerializer.fromMap(val, typeKey: typeKey))
+            ?.toList() ??
+        model.authors;
     return model;
   }