+++ test/serializer/various/various.g.dart
@@ -31,8 +31,8 @@ abstract class _$ModelIntSerializer implements Serializer<ModelInt> {
     if (model is! ModelInt) {
       model = createModel();
     }
-    model.bar = map["bar"];
-    model.clazzA = map["clazzA"];
+    model.bar = map["bar"] ?? model.bar;
+    model.clazzA = map["clazzA"] ?? model.clazzA;
     return model;
   }
 
@@ -68,8 +68,8 @@ abstract class _$ModelDoubleSerializer implements Serializer<ModelDouble> {
     if (model is! ModelDouble) {
       model = createModel();
     }
-    model.bar = map["bar"];
-    model.clazzA = map["clazzA"];
+    model.bar = map["bar"] ?? model.bar;
+    model.clazzA = map["clazzA"] ?? model.clazzA;
     return model;
   }
 
@@ -105,8 +105,8 @@ abstract class _$InheritanceSerializer implements Serializer<Inheritance> {
     if (model is! Inheritance) {
       model = createModel();
     }
-    model.clazzA = map["clazzA"];
-    model.clazzB = map["clazzB"];
+    model.clazzA = map["clazzA"] ?? model.clazzA;
+    model.clazzB = map["clazzB"] ?? model.clazzB;
     return model;
   }
 
@@ -144,8 +144,8 @@ abstract class _$DateSerializer implements Serializer<Date> {
     if (model is! Date) {
       model = createModel();
     }
-    model.date = dateDateTimeProcessor.deserialize(map["date"]);
-    model.clazzA = map["clazzA"];
+    model.date = dateDateTimeProcessor.deserialize(map["date"]) ?? model.date;
+    model.clazzA = map["clazzA"] ?? model.clazzA;
     return model;
   }
 
@@ -194,12 +194,14 @@ abstract class _$NullTestSerializer implements Serializer<NullTest> {
     if (model is! NullTest) {
       model = createModel();
     }
-    model.tests = map["tests"]?.map((String val) => val)?.toList();
-    model.test = map["test"];
+    model.tests =
+        map["tests"]?.map((String val) => val)?.toList() ?? model.tests;
+    model.test = map["test"] ?? model.test;
     model.testModel = map["testModel"]
-        ?.map(
-            (Map val) => fromModelIntSerializer.fromMap(val, typeKey: typeKey))
-        ?.toList();
+            ?.map((Map val) =>
+                fromModelIntSerializer.fromMap(val, typeKey: typeKey))
+            ?.toList() ??
+        model.testModel;
     return model;
   }
 
@@ -232,7 +234,7 @@ abstract class _$WithIgnoreSerializer implements Serializer<WithIgnore> {
     if (model is! WithIgnore) {
       model = createModel();
     }
-    model.a = map["a"];
+    model.a = map["a"] ?? model.a;
     return model;
   }
 
@@ -265,7 +267,7 @@ abstract class _$ModelRenamedSerializer implements Serializer<ModelRenamed> {
     if (model is! ModelRenamed) {
       model = createModel();
     }
-    model.original = map["renamed"];
+    model.original = map["renamed"] ?? model.original;
     return model;
   }
 
@@ -382,45 +384,57 @@ abstract class _$ComplexSerializer implements Serializer<Complex> {
     if (model is! Complex) {
       model = createModel();
     }
-    model.nums = map["nums"]?.map((num val) => val)?.toList();
-    model.strings = map["strings"]?.map((String val) => val)?.toList();
-    model.bools = map["bools"]?.map((bool val) => val)?.toList();
-    model.ints = map["ints"]?.map((int val) => val)?.toList();
-    model.doubles = map["doubles"]?.map((double val) => val)?.toList();
+    model.nums = map["nums"]?.map((num val) => val)?.toList() ?? model.nums;
+    model.strings =
+        map["strings"]?.map((String val) => val)?.toList() ?? model.strings;
+    model.bools = map["bools"]?.map((bool val) => val)?.toList() ?? model.bools;
+    model.ints = map["ints"]?.map((int val) => val)?.toList() ?? model.ints;
+    model.doubles =
+        map["doubles"]?.map((double val) => val)?.toList() ?? model.doubles;
     model.ignores = map["ignores"]
-        ?.map((Map val) =>
-            fromWithIgnoreSerializer.fromMap(val, typeKey: typeKey))
-        ?.toList();
+            ?.map((Map val) =>
+                fromWithIgnoreSerializer.fromMap(val, typeKey: typeKey))
+            ?.toList() ??
+        model.ignores;
     model.numSet =
         new MapMaker(map["numSet"], (String key) => key, (num value) {
-      return value;
-    }).model as dynamic;
+              return value;
+            }).model as dynamic ??
+            model.numSet;
     model.stringSet =
         new MapMaker(map["stringSet"], (String key) => key, (String value) {
-      return value;
-    }).model as dynamic;
+              return value;
+            }).model as dynamic ??
+            model.stringSet;
     model.boolSet =
         new MapMaker(map["boolSet"], (String key) => key, (bool value) {
-      return value;
-    }).model as dynamic;
+              return value;
+            }).model as dynamic ??
+            model.boolSet;
     model.intSet =
         new MapMaker(map["intSet"], (String key) => key, (int value) {
-      return value;
-    }).model as dynamic;
+              return value;
+            }).model as dynamic ??
+            model.intSet;
     model.doubleSet =
         new MapMaker(map["doubleSet"], (String key) => key, (double value) {
-      return value;
-    }).model as dynamic;
+              return value;
+            }).model as dynamic ??
+            model.doubleSet;
     model.ignoreSet =
         new MapMaker(map["ignoreSet"], (String key) => key, (Map value) {
-      return fromWithIgnoreSerializer.fromMap(value, typeKey: typeKey);
-    }).model as dynamic;
+              return fromWithIgnoreSerializer.fromMap(value, typeKey: typeKey);
+            }).model as dynamic ??
+            model.ignoreSet;
     model.listInnerMap1 = new MapMaker(
-        map["listInnerMap1"], (String key) => key, (List<String> value) {
-      return value?.map((String val) => val)?.toList();
-    }).model as dynamic;
-    model.dynamicMap = dynamicMapRawData.deserialize(map["dynamicMap"]);
-    model.dynamicList = dynamicListRawData.deserialize(map["dynamicList"]);
+            map["listInnerMap1"], (String key) => key, (List<String> value) {
+          return value?.map((String val) => val)?.toList();
+        }).model as dynamic ??
+        model.listInnerMap1;
+    model.dynamicMap =
+        dynamicMapRawData.deserialize(map["dynamicMap"]) ?? model.dynamicMap;
+    model.dynamicList =
+        dynamicListRawData.deserialize(map["dynamicList"]) ?? model.dynamicList;
     return model;
   }
 
@@ -454,7 +468,7 @@ abstract class _$CustomModelNameSerializer
     if (model is! CustomModelName) {
       model = createModel();
     }
-    model.foo = map["foo"];
+    model.foo = map["foo"] ?? model.foo;
     return model;
   }