+++ tool/travis.sh
@@ -9,16 +9,7 @@ chmod +x $(dirname -- "$0")/ensure_dartfmt.sh
 # dart bin/serializer.dart build
 
 # Run vm tests
-pub run test -p vm test/all_test.dart
-
-# Run browser tests
-if [ "${TRAVIS}" == "true" ]; then
-  export DISPLAY=:99.0
-  sh -e /etc/init.d/xvfb start
-fi
-pub serve test &
-while ! nc -z localhost 8080; do sleep 1; done; echo 'pub serve is up!'
-pub run test --pub-serve=8080 -p firefox test/all_browser_test.dart
+pub run test -p vm -p content-shell -p firefox -p chrome
 
 # Install dart_coveralls; gather and send coverage data.
 # Re activate coverall when a solution arrive